requirements.txt
# =============================================================================
# GlobeSpeak - AI-Powered Translation & Video Summarization Platform
# Requirements file for Python 3.12+ on Windows
# =============================================================================

# =============================================================================
# CORE DJANGO FRAMEWORK
# =============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
whitenoise==6.6.0

# =============================================================================
# DJANGO DEPENDENCIES
# =============================================================================
asgiref==3.9.1
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2

# =============================================================================
# TRANSLATION & LANGUAGE PROCESSING
# =============================================================================
deep-translator==1.11.4
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.62.3
proto-plus==1.26.1
protobuf==4.25.8

# =============================================================================
# COMPUTER VISION & IMAGE PROCESSING
# =============================================================================
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10

# =============================================================================
# SPEECH & AUDIO PROCESSING
# =============================================================================
speechrecognition==3.10.0
gtts==2.4.0
openai-whisper==20231117

# =============================================================================
# AI/ML & NATURAL LANGUAGE PROCESSING
# =============================================================================
transformers==4.35.2
torch==2.8.0
nltk==3.8.1
numpy==1.26.4
huggingface-hub==0.35.0
tokenizers==0.15.2
safetensors==0.6.2
tiktoken==0.11.0

# =============================================================================
# VIDEO PROCESSING & DOWNLOADING
# =============================================================================
yt-dlp==2025.9.5

# =============================================================================
# NUMERICAL COMPUTING & SCIENTIFIC LIBRARIES
# =============================================================================
numba==0.62.0
llvmlite==0.45.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.5

# =============================================================================
# WEB & HTTP LIBRARIES
# =============================================================================
requests==2.31.0
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# =============================================================================
# TEXT PROCESSING & PARSING
# =============================================================================
beautifulsoup4==4.13.5
soupsieve==2.8
regex==2025.9.18
PyYAML==6.0.2

# =============================================================================
# UTILITIES & TOOLS
# =============================================================================
python-dotenv==1.0.0
psutil==5.9.6
colorama==0.4.6
tqdm==4.67.1
click==8.1.8
packaging==25.0
filelock==3.19.1
more-itertools==10.8.0
fsspec==2025.9.0
joblib==1.5.2
setuptools==80.9.0

# =============================================================================
# TEMPLATE & UI LIBRARIES
# =============================================================================
Jinja2==3.1.6
MarkupSafe==3.0.2

# =============================================================================
# SECURITY & AUTHENTICATION
# =============================================================================
cachetools==5.5.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# =============================================================================
# TYPE HINTS & COMPATIBILITY
# =============================================================================
typing_extensions==4.15.0

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 1. Python Version: Requires Python 3.12+
# 2. Virtual Environment: Always use a virtual environment
# 3. FFmpeg: Install FFmpeg separately for video processing
#    - Download from: https://ffmpeg.org/download.html
#    - Add to PATH: C:\ffmpeg\bin
# 4. Tesseract: Install Tesseract OCR for image text extraction
#    - Download from: https://github.com/UB-Mannheim/tesseract/wiki
#    - Add to PATH: C:\Program Files\Tesseract-OCR
# 5. Installation Command:
#    pip install -r requirements.txt
# 6. For development, also install:
#    pip install django-debug-toolbar django-extensions

# =============================================================================
# OPTIONAL DEPENDENCIES (Uncomment if needed)
# =============================================================================
# redis==5.0.1  # For production Redis caching
# channels-redis==4.1.0  # For production WebSocket support
# celery==5.3.4  # For background task processing
# django-debug-toolbar==4.2.0  # For development debugging
# django-extensions==3.2.3  # For development utilities
# pytest==7.4.3  # For testing
# pytest-django==4.7.0  # For Django testing
# coverage==7.3.2  # For test coverage